+++ CHANGELOG.md
@@ -10,6 +10,9 @@ listed in the changelog.
 
 ## [Unreleased]
 
+### Changed
+
+- Updated `_auth` value in npm config to be scoped to the npm registry in Nexus [#668](https://github.com/opendevstack/ods-pipeline/issues/668)
 
 ## [0.10.0] - 2023-02-27